introduce delay between making requests
This is because the now-outdated servers (i.e. tagpro-radius) now simply point to the same tagpro-test location.

The server rotation functionality is now currently deprecated, but it may be useful again in the future if we are able to get profile data from the other servers (which isn't currently possible; i.e. tagpro-seattle.koalabeast.com/profile/<id> does not currently work).